smartreco/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pymongo>=4.3.0
chromadb>=0.4.0
groq>=0.3.0
langchain>=0.1.0
sentence-transformers>=2.2.2
requests>=2.28.0
python-dotenv>=1.0.0
pydantic>=1.10.0
